backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
anthropic
beautifulsoup4
pydantic-settings
openai
